Refactor processPhase3Message to remove processMessageCorePhase3
