+++ packages/core/characterisation-tests/helpers/processMessage.ts
@@ -4,14 +4,38 @@ import type Phase1Result from "../../phase1/types/Phase1Result"
 import type Phase2Result from "../../phase2/types/Phase2Result"
 import type Phase3Result from "../../phase3/types/Phase3Result"
 import type PncUpdateRequestError from "../../phase3/types/PncUpdateRequestError"
+import type { PncUpdateDataset } from "../../types/PncUpdateDataset"
 import type { ProcessMessageOptions } from "../types/ProcessMessageOptions"
 
 import CoreAuditLogger from "../../lib/auditLog/CoreAuditLogger"
+import { PncApiError } from "../../lib/pnc/PncGateway"
 import CorePhase1 from "../../phase1/phase1"
 import CorePhase2 from "../../phase2/phase2"
+import CorePhase3 from "../../phase3/phase3"
+import MockPncGateway from "../../tests/helpers/MockPncGateway"
 import parseIncomingMessage from "../../tests/helpers/parseIncomingMessage"
 import Phase from "../../types/Phase"
-import { createMockPncGateway, processMessageCorePhase3 } from "./processMessageCore"
+import generateMockPncQueryResult from "./generateMockPncQueryResult"
+
+const createMockPncGateway = (
+  messageXml: string,
+  {
+    recordable = true,
+    pncOverrides = {},
+    pncCaseType = "court",
+    pncErrorMessage,
+    pncMessage,
+    pncAdjudication = false
+  }: ProcessMessageOptions
+) => {
+  const mockPncResponse = recordable
+    ? pncErrorMessage
+      ? new PncApiError([pncErrorMessage])
+      : generateMockPncQueryResult(pncMessage ? pncMessage : messageXml, pncOverrides, pncCaseType, pncAdjudication)
+    : undefined
+
+  return new MockPncGateway(mockPncResponse)
+}
 
 export const processPhase1Message = async (
   messageXml: string,
@@ -31,8 +55,13 @@ export const processPhase2Message = (messageXml: string): Phase2Result => {
   return CorePhase2(incomingMessage, auditLogger)
 }
 
-export const processPhase3Message = (
+export const processPhase3Message = async (
   messageXml: string,
   options: ProcessMessageOptions = {}
-): Promise<Phase3Result | PncUpdateRequestError> =>
-  processMessageCorePhase3(messageXml, { phase: Phase.PHASE_3, ...options })
+): Promise<Phase3Result | PncUpdateRequestError> => {
+  const { message: incomingMessage } = parseIncomingMessage(messageXml)
+  const pncGateway = createMockPncGateway(messageXml, { phase: Phase.PHASE_3, ...options })
+  const auditLogger = new CoreAuditLogger(AuditLogEventSource.CorePhase3)
+
+  return await CorePhase3(incomingMessage as PncUpdateDataset, pncGateway, auditLogger)
+}
