--- packages/core/characterisation-tests/helpers/processMessageCore.ts
@@ -1,44 +1,0 @@
-import { AuditLogEventSource } from "@moj-bichard7/common/types/AuditLogEvent"
-
-import type Phase3Result from "../../phase3/types/Phase3Result"
-import type PncUpdateRequestError from "../../phase3/types/PncUpdateRequestError"
-import type { PncUpdateDataset } from "../../types/PncUpdateDataset"
-import type { ProcessMessageOptions } from "../types/ProcessMessageOptions"
-
-import CoreAuditLogger from "../../lib/auditLog/CoreAuditLogger"
-import { PncApiError } from "../../lib/pnc/PncGateway"
-import CorePhase3 from "../../phase3/phase3"
-import MockPncGateway from "../../tests/helpers/MockPncGateway"
-import parseIncomingMessage from "../../tests/helpers/parseIncomingMessage"
-import generateMockPncQueryResult from "./generateMockPncQueryResult"
-
-export const createMockPncGateway = (
-  messageXml: string,
-  {
-    recordable = true,
-    pncOverrides = {},
-    pncCaseType = "court",
-    pncErrorMessage,
-    pncMessage,
-    pncAdjudication = false
-  }: ProcessMessageOptions
-) => {
-  const mockPncResponse = recordable
-    ? pncErrorMessage
-      ? new PncApiError([pncErrorMessage])
-      : generateMockPncQueryResult(pncMessage ? pncMessage : messageXml, pncOverrides, pncCaseType, pncAdjudication)
-    : undefined
-
-  return new MockPncGateway(mockPncResponse)
-}
-
-export const processMessageCorePhase3 = async (
-  messageXml: string,
-  options: ProcessMessageOptions
-): Promise<Phase3Result | PncUpdateRequestError> => {
-  const { message: incomingMessage } = parseIncomingMessage(messageXml)
-  const pncGateway = createMockPncGateway(messageXml, options)
-  const auditLogger = new CoreAuditLogger(AuditLogEventSource.CorePhase3)
-
-  return await CorePhase3(incomingMessage as PncUpdateDataset, pncGateway, auditLogger)
-}
